# docs/api/swagger.yaml
basePath: /
definitions:
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.Code:
    enum:
    - INVALID_PAYLOAD
    - VALIDATION_FAILED
    - UNAUTHENTICATED
    - FORBIDDEN
    - RESOURCE_NOT_FOUND
    - CONFLICT
    - RATE_LIMITED
    - PAYLOAD_TOO_LARGE
    - SEARCH_UNAVAILABLE
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    type: string
    x-enum-varnames:
    - CodeInvalidPayload
    - CodeValidationFailed
    - CodeUnauthenticated
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeRateLimited
    - CodePayloadTooLarge
    - CodeSearchUnavailable
    - CodeInternal
    - CodeNotImplemented
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.InvalidParam:
    properties:
      code:
        type: string
      field:
        type: string
    type: object
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails:
    properties:
      code:
        $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.InvalidParam'
        type: array
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
    type: object
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_returns_domain.ItemCondition:
    enum:
    - unopened
    - opened
    - damaged
    type: string
    x-enum-varnames:
    - ItemConditionUnopened
    - ItemConditionOpened
    - ItemConditionDamaged
  github_com_VladHrytsaiuk_ecommerce-core_internal_returns_domain.RefundMode:
    enum:
    - full
    - partial
    - store_credit
    type: string
    x-enum-varnames:
    - RefundModeFull
    - RefundModePartial
    - RefundModeStoreCredit
...
  internal_returns_delivery_http.createRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_returns_delivery_http.itemRequest'
        maxItems: 50
        minItems: 1
        type: array
      order_id:
        type: string
      refund_mode:
        $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_returns_domain.RefundMode'
    required:
    - items
    - order_id
    type: object
  internal_returns_delivery_http.itemRequest:
    properties:
      condition:
        $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_returns_domain.ItemCondition'
      quantity:
        maximum: 10000
        minimum: 1
        type: integer
      reason:
        maxLength: 2000
        type: string
      variant_id:
        type: string
    required:
    - condition
    - quantity
    - variant_id
    type: object
  internal_returns_delivery_http.reasonRequest:
    properties:
      reason:
        maxLength: 2000
        type: string
    type: object
  internal_reviews_delivery_http.createRequest:
    properties:
      comment:
        type: string
      rating:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API for AquaWheel e-commerce project.
  title: AquaWheel Store API
  version: "1.0"
paths:
...
  /api/v1/admin/returns/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: Return request UUID
        in: path
        name: id
        required: true
        type: string
      - description: Approval reason
        in: body
        name: payload
        schema:
          $ref: '#/definitions/internal_returns_delivery_http.reasonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Approve a return request (v1 admin)
      tags:
      - Admin v1
  /api/v1/admin/returns/{id}/receive:
    post:
      consumes:
      - application/json
      parameters:
      - description: Return request UUID
        in: path
        name: id
        required: true
        type: string
      - description: Receipt reason
        in: body
        name: payload
        schema:
          $ref: '#/definitions/internal_returns_delivery_http.reasonRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Receive a return and schedule controlled settlement (v1 admin)
      tags:
      - Admin v1
...
  /api/v1/catalog/products/{product_id}/reviews:
    get:
      description: Returns only reviews an administrator has approved. Pending and
        rejected reviews are never exposed, including to their own author.
      parameters:
      - description: Product UUID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: List approved reviews for a product
      tags:
      - Catalog v1
    post:
      consumes:
      - application/json
      description: Requires authentication; the review is attributed to the caller
        and starts pending moderation. A customer may review a product once.
      parameters:
      - description: Product UUID
        in: path
        name: product_id
        required: true
        type: string
      - description: Rating and comment
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_reviews_delivery_http.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      security:
      - bearerAuth: []
      summary: Submit a product review
      tags:
      - Catalog v1
...
  /api/v1/customers/me/returns:
    post:
      consumes:
      - application/json
      parameters:
      - description: Return request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_returns_delivery_http.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Create a return request for the current customer (v1)
      tags:
      - Returns v1
...
securityDefinitions:
  bearerAuth:
    description: 'Введіть токен у форматі: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...